Change the text of intention
This is because Android Studio v3.0 has almost the same intentions. I
don't want the plugin users feel weird to have the similar name for the
intention. However, official intention seems not support cnverting the
layout depends on the cursor location. So, as a temporary solution, I
changed the intention name to prevent user confuse from the same
intentions.